Don't include Windows.h before datatype.h
It turns out that Windows.h defines a `TIME_MS` symbol which causes a
series of nearly impenetrable syntax errors when attemting to define the
`Datatype` enum class.
